Adding to HTML task
Adding a task that reaches out to calabash for processing.

Something is still VERY off with using calabash from within the plugin.
When you extend the calabash gradle task, you can't seem to set
the entity or URI resolvers without getting a class not found exception.

If you call javaexec indirectly, then it works as expected.

So the code behind the HTML task is really wonky. I'll be cleaning it up
soon once I can figure out what exactly is going on.